Moved common modules to kie-bc-editors-common
