+++ packages/kie-bc-editors-common/jest.config.js
