+++ packages/kie-bc-editors-common/src/GwtEditorMapping.ts
