+++ packages/kie-bc-editors-common/src/TextFormatter.ts
