+++ packages/kie-bc-editors-common/src/XmlFormatter.ts
