+++ packages/kie-bc-editors-common/src/consumedInteropApi/GwtAppFormerConsumedInteropApi.ts
