+++ packages/kie-bc-editors-common/src/exposedInteropApi/KeyboardShorcutsApi.ts
