+++ packages/kie-bc-editors-common/src/exposedInteropApi/ResourceContentExposedInteropApi.ts
