+++ packages/kie-bc-editors-common/src/exposedInteropApi/WorkspaceExposedInteropApi.ts
