+++ packages/kie-bc-editors-common/src/exposedInteropApi/index.ts
