+++ packages/kie-bc-editors-common/src/gwtStateControl/DefaultStateControlCommandRegistry.ts
