+++ packages/kie-bc-editors-common/src/i18n/KieBcEditorsI18n.ts
