+++ packages/kie-bc-editors-common/src/i18n/index.ts
