+++ packages/kie-bc-editors-common/src/i18n/locales/de.ts
