+++ packages/kie-bc-editors-common/src/i18n/locales/index.ts
