+++ packages/kie-bc-editors-common/src/i18n/setup.ts
