+++ packages/kie-bc-editors-common/src/index.ts
