+++ packages/kie-bc-editors-common/tests/GwtEditorWrapper.test.ts
@@ -17,12 +17,12 @@
  * under the License.
  */
 
-import { GwtEditorWrapper } from "@kie-tools/kie-bc-editors/dist/common/GwtEditorWrapper";
-import { GwtStateControlService } from "@kie-tools/kie-bc-editors/dist/common/gwtStateControl";
+import { GwtEditorWrapper } from "@kie-tools/kie-bc-editors-common/dist/GwtEditorWrapper";
+import { GwtStateControlService } from "@kie-tools/kie-bc-editors-common/dist/gwtStateControl";
 import { EditorTheme, KogitoEditorChannelApi } from "@kie-tools-core/editor/dist/api";
 import { messageBusClientApiMock } from "@kie-tools-core/envelope-bus/dist-tests/messageBusClientApiMock";
 import { I18n } from "@kie-tools-core/i18n/dist/core";
-import { kieBcEditorsI18nDefaults, kieBcEditorsI18nDictionaries } from "@kie-tools/kie-bc-editors/dist/common/i18n";
+import { kieBcEditorsI18nDefaults, kieBcEditorsI18nDictionaries } from "@kie-tools/kie-bc-editors-common/dist/i18n";
 
 const MockEditor = jest.fn(() => ({
   undo: jest.fn(),
