+++ packages/kie-bc-editors-common/tests/__mocks__/styleMock.js
