+++ packages/kie-bc-editors-common/tests/gwtStateControl/GwtCommandRegistry.test.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { DefaultStateControlCommandRegistry } from "@kie-tools/kie-bc-editors/dist/common/gwtStateControl";
+import { DefaultStateControlCommandRegistry } from "@kie-tools/kie-bc-editors-common/dist/gwtStateControl";
 import { KogitoEditorChannelApi } from "@kie-tools-core/editor/dist/api";
 import { MessageBusClientApi } from "@kie-tools-core/envelope-bus/dist/api";
 import { messageBusClientApiMock } from "@kie-tools-core/envelope-bus/dist-tests/messageBusClientApiMock";
