+++ packages/kie-bc-editors-common/tests/jest.setup.ts
