+++ packages/kie-bc-editors-common/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@kie-tools/tsconfig/tsconfig.cjs.json",
+  "compilerOptions": {
+    "jsx": "react-jsx"
+  }
+}
