+++ packages/kie-bc-editors-common/tsconfig.tests.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "types": ["@testing-library/jest-dom"]
+  }
+}
