+++ packages/kie-bc-editors/src/bpmn/api/BpmnEditorLanguageData.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { editors, GwtLanguageData } from "../../common";
+import { editors, GwtLanguageData } from "@kie-tools/kie-bc-editors-common";
 
 export function getBpmnLanguageData(resourcesPathPrefix: string): GwtLanguageData {
   return {
