+++ packages/kie-bc-editors/src/scesim/api/SceSimEditorLanguageData.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { editors, GwtLanguageData } from "../../common";
+import { editors, GwtLanguageData } from "@kie-tools/kie-bc-editors-common";
 
 export function getSceSimLanguageData(resourcesPathPrefix: string): GwtLanguageData {
   return {
