+++ packages/serverless-workflow-diagram-editor-envelope/src/envelope/ServerlessWorkflowDiagramEditor.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { GwtEditorWrapper } from "@kie-tools/kie-bc-editors/dist/common";
+import { GwtEditorWrapper } from "@kie-tools/kie-bc-editors-common/dist";
 import { StunnerCanvas, StunnerEditor, StunnerNode, StunnerSession } from "../api/StunnerAPI";
 
 interface CustomWindow extends Window {
