+++ packages/serverless-workflow-diagram-editor-envelope/src/envelope/ServerlessWorkflowDiagramEditorFactory.ts
@@ -23,7 +23,7 @@ import {
   KogitoEditorEnvelopeContextType,
   EditorTheme,
 } from "@kie-tools-core/editor/dist/api";
-import { GwtEditorWrapperFactory } from "@kie-tools/kie-bc-editors/dist/common";
+import { GwtEditorWrapperFactory } from "@kie-tools/kie-bc-editors-common/dist";
 import {
   getServerlessWorkflowLanguageData,
   ServerlessWorkflowDiagramEditorChannelApi,
