+++ repo/graph.dot
@@ -138,6 +138,7 @@ digraph G {
   "@kie-tools/runtime-tools-process-dev-ui-webapp" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "@kie-tools/yaml-language-server" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "@kie-tools-core/k8s-yaml-to-apiserver-requests" [ color = "purple", fontcolor = "purple", style = "rounded" ];
+  "@kie-tools/kie-bc-editors-common" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "kie-editors-dev-vscode-extension" [ color = "black", fontcolor = "black", style = "dashed, rounded" ];
   "@kie-tools/pmml-editor" [ color = "blue", fontcolor = "blue", style = "rounded" ];
   "@kie-tools/vscode-extension-common-test-helpers" [ color = "blue", fontcolor = "blue", style = "rounded" ];
@@ -448,10 +449,11 @@ digraph G {
   "@kie-tools-core/k8s-yaml-to-apiserver-requests" -> "@kie-tools/cors-proxy-api" [ style = "dashed", color = "purple" ];
   "@kie-tools-core/keyboard-shortcuts" -> "@kie-tools-core/envelope-bus" [ style = "solid", color = "purple" ];
   "@kie-tools-core/keyboard-shortcuts" -> "@kie-tools-core/operating-system" [ style = "solid", color = "purple" ];
-  "@kie-tools/kie-bc-editors" -> "@kie-tools-core/editor" [ style = "solid", color = "blue" ];
-  "@kie-tools/kie-bc-editors" -> "@kie-tools-core/vscode-java-code-completion" [ style = "solid", color = "blue" ];
   "@kie-tools/kie-bc-editors" -> "@kie-tools/dmn-language-service" [ style = "solid", color = "blue" ];
+  "@kie-tools/kie-bc-editors" -> "@kie-tools/kie-bc-editors-common" [ style = "solid", color = "blue" ];
   "@kie-tools/kie-bc-editors" -> "@kie-tools/pmml-editor-marshaller" [ style = "solid", color = "blue" ];
+  "@kie-tools/kie-bc-editors-common" -> "@kie-tools-core/editor" [ style = "solid", color = "blue" ];
+  "@kie-tools/kie-bc-editors-common" -> "@kie-tools-core/vscode-java-code-completion" [ style = "solid", color = "blue" ];
   "kie-editors-dev-vscode-extension" -> "@kie-tools-core/vscode-extension" [ style = "solid", color = "black" ];
   "kie-editors-dev-vscode-extension" -> "@kie-tools/form-code-generator-vscode-command" [ style = "solid", color = "black" ];
   "kie-editors-dev-vscode-extension" -> "@kie-tools/kie-bc-editors" [ style = "solid", color = "black" ];
@@ -611,7 +613,7 @@ digraph G {
   "@kie-tools/serverless-workflow-diagram-editor" -> "@kie-tools/maven-base" [ style = "solid", color = "black" ];
   "@kie-tools/serverless-workflow-diagram-editor-assets" -> "@kie-tools/serverless-workflow-diagram-editor" [ style = "dashed", color = "blue" ];
   "@kie-tools/serverless-workflow-diagram-editor-assets" -> "@kie-tools/serverless-workflow-diagram-editor-envelope" [ style = "dashed", color = "blue" ];
-  "@kie-tools/serverless-workflow-diagram-editor-envelope" -> "@kie-tools/kie-bc-editors" [ style = "solid", color = "blue" ];
+  "@kie-tools/serverless-workflow-diagram-editor-envelope" -> "@kie-tools/kie-bc-editors-common" [ style = "solid", color = "blue" ];
   "@kie-tools/serverless-workflow-jq-expressions" -> "@kie-tools/serverless-workflow-service-catalog" [ style = "solid", color = "blue" ];
   "@kie-tools/serverless-workflow-language-service" -> "@kie-tools/json-yaml-language-service" [ style = "solid", color = "blue" ];
   "@kie-tools/serverless-workflow-language-service" -> "@kie-tools/serverless-workflow-jq-expressions" [ style = "solid", color = "blue" ];
