+++ repo/graph.json
@@ -17,14 +17,14 @@
       { "id": "@kie-tools/stunner-editors" },
       { "id": "@kie-tools/eslint" },
       { "id": "@kie-tools/root-env" },
-      { "id": "@kie-tools-core/editor" },
-      { "id": "@kie-tools-core/vscode-java-code-completion" },
       { "id": "@kie-tools/dmn-language-service" },
+      { "id": "@kie-tools/kie-bc-editors-common" },
       { "id": "@kie-tools/pmml-editor-marshaller" },
       { "id": "@kie-tools/stunner-editors-dmn-loader" },
       { "id": "@kie-tools-examples/dmn-editor-classic-on-webapp" },
       { "id": "@kie-tools-examples/dmn-editor-on-webapp" },
       { "id": "@kie-tools/dmn-editor-envelope" },
+      { "id": "@kie-tools-core/editor" },
       { "id": "@kie-tools/dmn-editor" },
       { "id": "@kie-tools-examples/dmn-editor-standalone-on-webapp" },
       { "id": "@kie-tools/dmn-editor-standalone" },
@@ -50,6 +50,7 @@
         "id": "kie-tools-examples-micro-frontends-multiplying-architecture-base64png-editor-vscode-extension"
       },
       { "id": "@kie-tools-core/vscode-extension" },
+      { "id": "@kie-tools-core/vscode-java-code-completion" },
       {
         "id": "@kie-tools-examples/micro-frontends-multiplying-architecture-ping-pong-view"
       },
@@ -332,17 +333,12 @@
       },
       {
         "source": "@kie-tools/kie-bc-editors",
-        "target": "@kie-tools-core/editor",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools/kie-bc-editors",
-        "target": "@kie-tools-core/vscode-java-code-completion",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools/kie-bc-editors",
         "target": "@kie-tools/dmn-language-service",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/kie-bc-editors",
+        "target": "@kie-tools/kie-bc-editors-common",
         "weight": 1
       },
       {
@@ -356,41 +352,6 @@
         "weight": 1
       },
       {
-        "source": "@kie-tools-core/editor",
-        "target": "@kie-tools-core/envelope",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/editor",
-        "target": "@kie-tools-core/keyboard-shortcuts",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/editor",
-        "target": "@kie-tools-core/notifications",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/editor",
-        "target": "@kie-tools-core/patternfly-base",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/vscode-java-code-completion",
-        "target": "@kie-tools/eslint",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/vscode-java-code-completion",
-        "target": "@kie-tools/root-env",
-        "weight": 1
-      },
-      {
-        "source": "@kie-tools-core/vscode-java-code-completion",
-        "target": "@kie-tools/tsconfig",
-        "weight": 1
-      },
-      {
         "source": "@kie-tools/dmn-language-service",
         "target": "@kie-tools/dmn-feel-antlr4-parser",
         "weight": 1
@@ -398,6 +359,16 @@
       {
         "source": "@kie-tools/dmn-language-service",
         "target": "@kie-tools-core/switch-expression-ts",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/kie-bc-editors-common",
+        "target": "@kie-tools-core/editor",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools/kie-bc-editors-common",
+        "target": "@kie-tools-core/vscode-java-code-completion",
         "weight": 1
       },
       {
@@ -453,6 +424,26 @@
       {
         "source": "@kie-tools/dmn-editor-envelope",
         "target": "@kie-tools/dmn-editor",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/editor",
+        "target": "@kie-tools-core/envelope",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/editor",
+        "target": "@kie-tools-core/keyboard-shortcuts",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/editor",
+        "target": "@kie-tools-core/notifications",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/editor",
+        "target": "@kie-tools-core/patternfly-base",
         "weight": 1
       },
       {
@@ -588,6 +579,21 @@
       {
         "source": "@kie-tools-core/vscode-extension",
         "target": "@kie-tools-core/vscode-java-code-completion",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/vscode-java-code-completion",
+        "target": "@kie-tools/eslint",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/vscode-java-code-completion",
+        "target": "@kie-tools/root-env",
+        "weight": 1
+      },
+      {
+        "source": "@kie-tools-core/vscode-java-code-completion",
+        "target": "@kie-tools/tsconfig",
         "weight": 1
       },
       {
@@ -2577,7 +2583,7 @@
       },
       {
         "source": "@kie-tools/serverless-workflow-diagram-editor-envelope",
-        "target": "@kie-tools/kie-bc-editors",
+        "target": "@kie-tools/kie-bc-editors-common",
         "weight": 1
       },
       {
@@ -2897,6 +2903,7 @@
     ["@kie-tools-core/k8s-yaml-to-apiserver-requests", "packages/k8s-yaml-to-apiserver-requests"],
     ["@kie-tools-core/keyboard-shortcuts", "packages/keyboard-shortcuts"],
     ["@kie-tools/kie-bc-editors", "packages/kie-bc-editors"],
+    ["@kie-tools/kie-bc-editors-common", "packages/kie-bc-editors-common"],
     ["kie-editors-dev-vscode-extension", "packages/kie-editors-dev-vscode-extension"],
     ["@kie-tools/kie-editors-standalone", "packages/kie-editors-standalone"],
     ["@kie-tools/kie-sandbox-accelerator-quarkus", "packages/kie-sandbox-accelerator-quarkus"],
